requirements.txt
requests>=2.31
pyyaml>=6.0
tqdm>=4.66
pytest>=8.0
ruff>=0.5